requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit>=1.28.0
pandas>=2.0.0
matplotlib>=3.7.0
pillow>=10.0.0
